requirements.txt
gradio
google-adk
google-generativeai
anthropic
deep_translator 
streamlit 
